% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_barcodes_scars.r
\name{extract_barcodes}
\alias{extract_barcodes}
\title{Extract iTracer barcodes}
\usage{
extract_barcodes(file, yield_size = NA, min_reads = 3, library = NA)
}
\arguments{
\item{file}{BAM file}

\item{yield_size}{Chunk size when reading the BAM file. Default is to read in the whole file}

\item{min_reads}{Minimal read number for a UMI to be considered as valid}

\item{library}{Label of the library}
}
\description{
Extract iTracer barcode from the BAM file of iTracer barcode library to the barcode reference with CellRanger
}
