% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_barcodes_scars.r
\name{extract_scars}
\alias{extract_scars}
\title{Extract iTracer scars}
\usage{
extract_scars(
  file,
  yield_size = NA,
  target = c("Scar", "Scar_Tomato"),
  min_reads = 3,
  library = NA
)
}
\arguments{
\item{file}{BAM file}

\item{yield_size}{Chunk size when reading the BAM file. Default is to read in the whole file}

\item{target}{Names of the valid target sequences in the reference}

\item{min_reads}{Minimal read number for a UMI to be considered as valid}

\item{library}{Label of the library}
}
\description{
Extract iTracer scars from the BAM file of iTracer scar library to the scar reference with CellRanger
}
