% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_merge_barcodes_scars.r
\name{incorporate_to_seurat}
\alias{incorporate_to_seurat}
\title{Incorporate the iTracer readout to the Seurat object}
\usage{
incorporate_to_seurat(
  seurat,
  df_iTracer,
  col_barcode = "GeneBarcodeMerged",
  col_scar = "ScarMerged"
)
}
\arguments{
\item{seurat}{Seurat object}

\item{df_iTracer}{iTracer readout data frame}

\item{col_barcode}{Column name in the seurat object meta.data slot for the iTracer barcodes. Default is 'GeneBarcodeMerged'}

\item{col_scar}{Column name in the seurat object meta.data slot for the iTracer scars. Default is 'ScarMerged'}
}
\description{
Incorporate the iTracer readout to the corresponding
Seurat object
}
