% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_merge_barcodes_scars.r
\name{merge_barcode_scars}
\alias{merge_barcode_scars}
\title{Merge the filtered iTracer barcode and scar data frames}
\usage{
merge_barcode_scars(df_barcode, df_scar)
}
\arguments{
\item{df_barcode}{Barcode data frame}

\item{df_scar}{Scar data frame}
}
\description{
Merge the filtered iTracer barcode and scar data frames, by considering only those presented
at the same UMIs in the same cells
}
