% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.r
\name{plot_tree}
\alias{plot_tree}
\title{Plot the lineage tree}
\usage{
plot_tree(
  metadata_df,
  col_lab = NA,
  col_barcode = "GeneBarcodeMerged",
  col_scar = "ScarMerged",
  edge_alpha = 0.25,
  edge_width = 0.1
)
}
\arguments{
\item{metadata_df}{Metadata data frame}

\item{col_lab}{The metadata column name of cell annotations}

\item{col_barcode}{The metadata column name of iTracer barcodes}

\item{col_scar}{The metadata column name of iTracer scars}

\item{edge_alpha}{Edge alpha of the tree}

\item{edge_width}{Edge width of the tree}
}
\description{
Plot the lineage tree in the wheel shape
}
